Add multibench support to all-in-one json
Add benchmarks block to identify bench ids. The blockbreaker
utility extracts the benchmark from the runfile and processes
multibench config.

The mv-params for multibench is extracted from the 'benchmarks'
block. To keep backwards compatibility with the single bench mode,
it is still possible to define the mv-params block from the top
level of the JSON file.

For both single bench and multibench on the new schema, use
blockbreaker w/ --config mv-params and --benchmark <benchmark name>.
For single bench on the old schema, use --config mv-params and
--index <index number>.

This requires additional changes on the crucible code to fully
support 'benchmarks' block from the json runfile, and then
support it from the crucible cli.